organize wallet components
