+++ src/app/wallet/page.tsx
@@ -1,4 +1,4 @@
-import { WalletContainer } from './wallet-container';
+import { WalletContainer } from './components/wallet-container';
 
 export default function Wallet() {
   return <WalletContainer />;
